+++ android/src/main/kotlin/com/sharmadhiraj/installed_apps/Util.kt
@@ -4,14 +4,9 @@ import android.content.Context
 import android.content.pm.ApplicationInfo
 import android.content.pm.PackageInfo
 import android.content.pm.PackageManager
-import android.graphics.Bitmap
-import android.graphics.Canvas
-import android.graphics.drawable.BitmapDrawable
-import android.graphics.drawable.Drawable
 import android.os.Build.VERSION.SDK_INT
-import android.os.Build.VERSION_CODES.N_MR1
 import android.os.Build.VERSION_CODES.P
-import java.io.ByteArrayOutputStream
+import java.io.File
 
 class Util {
 
@@ -26,34 +21,16 @@ class Util {
             map["name"] = packageManager.getApplicationLabel(app)
             map["package_name"] = app.packageName
             map["icon"] =
-                if (withIcon) drawableToByteArray(app.loadIcon(packageManager)) else ByteArray(0)
+                if (withIcon) DrawableUtil.drawableToByteArray(app.loadIcon(packageManager))
+                else ByteArray(0)
             val packageInfo = packageManager.getPackageInfo(app.packageName, 0)
             map["version_name"] = packageInfo.versionName
             map["version_code"] = getVersionCode(packageInfo)
+            map["built_with"] = BuiltWithUtil.getPlatform(packageInfo.applicationInfo)
+            map["installed_timestamp"] = File(packageInfo.applicationInfo.sourceDir).lastModified()
             return map
         }
 
-        private fun drawableToByteArray(drawable: Drawable): ByteArray {
-            val bitmap = drawableToBitmap(drawable)
-            val stream = ByteArrayOutputStream()
-            bitmap.compress(Bitmap.CompressFormat.PNG, 100, stream)
-            return stream.toByteArray()
-        }
-
-        private fun drawableToBitmap(drawable: Drawable): Bitmap {
-            if (drawable is BitmapDrawable) {
-                return drawable.bitmap
-            }
-            val bitmap = Bitmap.createBitmap(
-                drawable.intrinsicWidth,
-                drawable.intrinsicHeight,
-                Bitmap.Config.ARGB_8888
-            )
-            val canvas = Canvas(bitmap)
-            drawable.setBounds(0, 0, canvas.width, canvas.height)
-            drawable.draw(canvas)
-            return bitmap
-        }
 
         fun getPackageManager(context: Context): PackageManager {
             return context.packageManager