+++ CHANGELOG.md
@@ -1,20 +1,36 @@
+## 1.5.0
+
+* Include built with and installed timestamp
+* Internal improvements
+
+## 1.4.0
+
+* Fixed issues (13,14) reported on GitHub
+* Implemented uninstallApp & isAppInstalled method
+
 ## 1.3.1
+
 * Fixed BinaryMessenger Nullability
 
 ## 1.3.0
+
 * Android V2 embedding migration
-* Fix UI freeze issue during getting instgalled apps
+* Fix UI freeze issue during getting installed apps
 
 ## 1.2.0
+
 * Null Safety migration
 
 ## 1.1.1
+
 * Add getAppInfo() & isSystemApp() method.
 * Refactor existing methods
 
 ## 1.0.0
+
 * Fixes & Optimization
 
 ## 0.0.1
+
 * Get list of installed app
 * Start app with package name